move network drawers to views folders
