+++ packages/desktop/components/ContactAddressCard.svelte
@@ -16,7 +16,7 @@
     import { openPopup, PopupId } from '@desktop/auxiliary/popup'
     import { closeDrawer } from '@desktop/auxiliary/drawer'
 
-    import { ContactBookRoute } from '@views/dashboard/contact-book/contact-book-route.enum'
+    import { ContactBookRoute } from '@views/dashboard/drawers/contact-book/contact-book-route.enum'
     import { SendFlowRouter, sendFlowRouter } from '@views/dashboard/send-flow'
     import { localize } from '@core/i18n'
