+++ packages/desktop/components/drawers/index.ts
@@ -1,4 +1,1 @@
-export { default as Drawer } from './Drawer.svelte'
 export { default as DrawerTemplate } from './DrawerTemplate.svelte'
-
-export { default as NetworkConfigDrawerRouter } from './network-config/NetworkConfigDrawerRouter.svelte'
