--- packages/desktop/lib/routers/enums/index.ts
@@ -1,1 +1,0 @@
-export * from './network-config-route.enum'
