+++ packages/desktop/lib/routers/index.ts
@@ -1,6 +1,3 @@
 export * from './actions'
-export * from './enums'
 export * from './interfaces'
-export * from './routers'
-export * from './types'
 export * from './utils'
