--- packages/desktop/lib/routers/routers/index.ts
@@ -1,1 +1,0 @@
-export * from './network-config.router'
