--- packages/desktop/lib/routers/types/drawer-route.type.ts
@@ -1,4 +1,0 @@
-import { NetworkConfigRoute } from '../enums'
-
-// TODO: this is wrong
-export type DrawerRoute = NetworkConfigRoute
