--- packages/desktop/lib/routers/types/index.ts
@@ -1,1 +1,0 @@
-export * from './drawer-route.type'
