+++ packages/desktop/views/dashboard/Dashboard.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import { Drawer } from '@components'
+    import Drawer from './Drawer.svelte'
     import { selectedAccount, selectedAccountIndex } from '@core/account/stores'
     import { Platform } from '@core/app'
     import { clearLayer2TokensPoll, pollLayer2Tokens } from '@core/layer-2'
