+++ packages/desktop/views/dashboard/Drawer.svelte
@@ -1,20 +1,13 @@
 <script lang="ts">
-    import { fade, fly } from 'svelte/transition'
-    import { NetworkConfigDrawerRouter } from '@components'
-    import { Router } from '@core/router'
-    import { closeDrawer, DrawerDirection, DrawerId, drawerState } from '@desktop/auxiliary/drawer'
-    import { DrawerRoute } from '@desktop/routers'
     import { Icon as IconEnum } from '@auxiliary/icon'
+    import { DrawerDirection, DrawerId, closeDrawer, drawerState } from '@desktop/auxiliary/drawer'
     import { Icon } from '@ui'
-    import { ContactBookRouterView } from '@views/dashboard/contact-book'
-    import { DappsRouterView } from '@views/dashboard/dapps'
+    import { ContactBookRouterView, NetworkConfigDrawerRouter, DappsRouterView } from './drawers'
+    import { fade, fly } from 'svelte/transition'
 
     export let onClose: () => unknown = () => {}
 
     const DRAWER_ANIMATION_DURATION_MS = 200
-
-    let drawerRoute: DrawerRoute
-    let drawerRouter: Router<DrawerRoute>
 
     let direction: { x: number; y: number }
     let position: string
@@ -72,7 +65,7 @@
         >
             <div class="flex flex-col h-full">
                 {#if $drawerState.id === DrawerId.NetworkConfig}
-                    <NetworkConfigDrawerRouter bind:drawerRoute bind:drawerRouter />
+                    <NetworkConfigDrawerRouter />
                 {:else if $drawerState.id === DrawerId.Dapps}
                     <DappsRouterView />
                 {:else if $drawerState.id === DrawerId.ContactBook}
