+++ packages/desktop/views/dashboard/drawers/contact-book/drawers/ContactListDrawer.svelte
