+++ packages/desktop/views/dashboard/drawers/contact-book/drawers/RemoveContactDrawer.svelte
