+++ packages/desktop/views/dashboard/drawers/contact-book/drawers/RemoveNetworkAddressesDrawer.svelte
