+++ packages/desktop/views/dashboard/drawers/dapps/DappsRouterView.svelte
