+++ packages/desktop/views/dashboard/drawers/dapps/drawers/ConnectedDappsDrawer.svelte
