+++ packages/desktop/views/dashboard/drawers/network-drawer/index.ts
@@ -1,0 +1,4 @@
+export { default as NetworkConfigDrawerRouter } from './NetworkConfigDrawerRouter.svelte'
+
+export * from './network-config-route.enum'
+export * from './network-config.router'
