+++ packages/desktop/views/dashboard/drawers/network-drawer/network-config-route.enum.ts
