+++ packages/desktop/views/dashboard/drawers/network-drawer/network-config.router.ts
@@ -1,8 +1,6 @@
 import { get, writable } from 'svelte/store'
-
 import { Router } from '@core/router'
-
-import { NetworkConfigRoute } from '../enums'
+import { NetworkConfigRoute } from './network-config-route.enum'
 
 export const networkConfigRoute = writable<NetworkConfigRoute>(null)
 export const networkConfigRouter = writable<NetworkConfigRouter>(null)
