+++ packages/desktop/views/dashboard/drawers/network-drawer/views/AddChainDrawer.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
-    import { localize } from '@core/i18n'
     import { DrawerTemplate } from '@components'
-    import { AddIscpChainForm } from './components'
+    import { localize } from '@core/i18n'
     import { Router } from '@core/router'
-    import { DrawerRoute } from '@desktop/routers'
+    import { NetworkConfigRoute } from '../'
+    import { AddIscpChainForm } from './components'
 
-    export let drawerRouter: Router<DrawerRoute>
+    export let drawerRouter: Router<NetworkConfigRoute>
 </script>
 
 <DrawerTemplate title={localize('views.dashboard.drawers.networkConfig.addChain.title')} {drawerRouter}>
