+++ packages/desktop/views/dashboard/drawers/network-drawer/views/ConnectLedgerDeviceDrawer.svelte
@@ -9,10 +9,10 @@
         ledgerNanoStatus,
     } from '@core/ledger'
     import { Router } from '@core/router'
-    import { DrawerRoute, NetworkConfigRoute, networkConfigRouter } from '@desktop/routers'
     import { Animation, FontWeight, Icon, Pane, Text, TextType } from '@ui'
+    import { NetworkConfigRoute, networkConfigRouter } from '../'
 
-    export let drawerRouter: Router<DrawerRoute>
+    export let drawerRouter: Router<NetworkConfigRoute>
 
     const LOCALE_BASE_PATH = 'views.dashboard.drawers.networkConfig.connectLedgerDevice'
