+++ packages/desktop/views/dashboard/drawers/network-drawer/views/components/EvmChainInformation.svelte
