+++ packages/desktop/views/dashboard/drawers/network-drawer/views/components/IscpChainInformation.svelte
