+++ packages/desktop/views/dashboard/drawers/network-drawer/views/components/index.ts
