+++ packages/desktop/views/dashboard/drawers/network-drawer/views/index.ts
